Fix Windows tests falsely passing even if one test has failed
Description: It turns out that adding the failed test count to failedTestCount in ctest did not work. I suspect that at some point this has worked, and suspect that it got broken when switching to 2015. I did not dig into that though and me and Jelani just went ahead and fixed it by adding some macro magic in ctest.